+++ packages/terra-core-docs/src/terra-dev-site/doc/action-header/example/ActionHeaderWithDecorativeIcon.jsx
@@ -1,0 +1,53 @@
+import React, { useState } from 'react';
+import Button, { IconButton, IconTypes } from 'terra-button';
+import ActionHeader from 'terra-action-header';
+import ContentContainer from 'terra-content-container';
+import Card from 'terra-card';
+import DecoIconPrinter from 'terra-icon/lib/icon/decorative/IconPrinter';
+
+const ActionHeaderWithDecorativeIcon = () => {
+  const text = ['Current Patient: Alex',
+    ' Patient age: 25', ' Patient Id: 11403'];
+  const [isOpen, setOpen] = useState(false);
+  const DecorativeIcon = <DecoIconPrinter />;
+
+  const handleOpenModal = () => {
+    setOpen(true);
+  };
+
+  const handleCloseModal = () => {
+    setOpen(false);
+  };
+
+  if (isOpen) {
+    return (
+      <div>
+        <Card>
+          <Card.Body>
+            <ContentContainer
+              header={(
+                <ActionHeader text="Patient Details" level={2} onBack={handleCloseModal}>
+                  <IconButton icon={DecorativeIcon} iconType={IconTypes.DECORATIVE} text="Print Patient Details" />
+                </ActionHeader>
+)}
+            >
+              <p>{text}</p>
+            </ContentContainer>
+          </Card.Body>
+        </Card>
+      </div>
+    );
+  }
+
+  return (
+    <Card>
+      <div>
+        <Card.Body>
+          <h3>Patient Details</h3>
+          <Button id="trigger-dialog-modal" text="Show Patient Details" onClick={handleOpenModal} />
+        </Card.Body>
+      </div>
+    </Card>
+  );
+};
+export default ActionHeaderWithDecorativeIcon;