+++ packages/terra-core-docs/CHANGELOG.md
@@ -22,6 +22,7 @@
   * Added Accessibility guide for `terra-button`.
   * Changes to Accessibility guide for `terra-button`.
   * Added a11y changes to `terra-action-header` examples.
+  * Added Accessibility guide for `terra-action-header`.
 
 ## 1.14.1 - (July 14, 2022)